requirements.txt
asgiref
certifi
colorama
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
filelock
iniconfig
Markdown
MouseInfo
mss
numpy
opencv-contrib-python
packaging
Pillow
pipenv
platformdirs
pluggy
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytest
pytweening
pytz
sqlparse
typing_extensions
tzdata
virtualenv